+++ src/ebird/checklists/models/__init__.py
@@ -3,6 +3,15 @@ from .country import Country
 from .location import Location
 from .observation import Observation
 from .observer import Observer
+from .region import Region
 from .species import Species
 
-__all__ = ("Checklist", "Country", "Location", "Observation", "Observer", "Species")
+__all__ = (
+    "Checklist",
+    "Country",
+    "Location",
+    "Observation",
+    "Observer",
+    "Region",
+    "Species",
+)